MarketPulse/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
mcp>=1.0.0
python-dotenv>=1.0.0
firecrawl-py>=0.0.16
asyncio-mqtt>=0.11.0